requirements.txt
Pillow>=10.0.0
requests>=2.31.0
tkinter-tooltip>=2.1.0